Release/1.2.0 (#4)
* Fix: loaded poi's on map loaded

Added in the show/hide special points of interest when the map is finished loading in addition to when the zoom level is changed

* Disable panTo for user location

* Fix for older browsers

After some testing on one of the older mobile devices, I've introduced some "use strict" clauses and corrected a few typos to ensure they work properly

* Updated map tiles to uncompressed, but newer, versions

* UX feedback fixes

Fixed a few things that were broken, changed some things due to UX dept. feedback:

- Corrected map positioning of POI items so they display in centre of coordinate
- Corrected names of .svg files so the map marker pulls through
- Changed the popup to remove the long description if it doesn't have one
- Added a zoomlimit check for higher zooms on larger devices
- Wrapped the global notice checker in an if statement to check if we have a global notice to display

* Geolocation error supression

Shut off the alert generated by the geolocation stuff as it was blocking the rest of the script from working.

Also implemented a softer, non-blocking error using our PL alert system

* Changed loading of map interaction buttons

Moved the campus buttons and feedback buttons into a hidden state until the map tiles have loaded. It looks less broken this way.

* Remaining autocomplete items causing error

the 'empty()' call on the $autocompletelist variable was throwing errors because this has been abstracted out of the library now and into the PL

* Custom marker ordering

-Changed the Google Maps layer where the custom points of interest are rendered to show BELOW the current toggleable markers
- Removed some redundent html/js
- Swapped out the Typkit call for the default stylesheet link/rel version

* Updated release notes

* Abstracted map title

Quick change to allow for the editing of the campus map title using the mapconfig.json file instead of hard coding per application.

* Updated static files to the latest pattern library release

* Refactoring as per Codacy review

* Further Codacity refactorings

* Codacity refactoring

* Refactoring work

* Updated POI coords

Paul updated the POI coordinates so that the labels display over the map tile names, thus improving the look.