Fix subpath in manifest.json
When adding the map to the home screen (tested on Android), tapping the tile launches https://york.ac.uk, not the map. This PR sets the `start_url` of the manifest properly. It also sets the navigation scope, so going to pages outside the map will open the default browser.